requirements.txt
torch
torchvision
datasets
Pillow
numpy
